🚚 Move: 아이콘 경로 변경
